Fixed sortlist and added Type as stringname in sortsettings

+++ BorjesLIA/BorjesLIA/Models/Diesel/DieselQuarterPriceModel.cs
@@ -48,7 +48,7 @@ namespace BorjesLIA.Models.Diesel
         public int PlacingOrder { get; set; }
 
         [Display(Name = "Typ")]
-        public decimal Type { get; set; }
+        public int Type { get; set; }
 
         [Display(Name = "Aktiv")]
         public bool Active { get; set; }
